Refactor endpoint expression helpers
